+++ examples/resources/ise_sxp_vpn/resource.tf
@@ -1,3 +1,3 @@
 resource "ise_sxp_vpn" "example" {
-  sxp_vpn_name = "SXPVpn1"
+  sxp_vpn_name = "sxpvpn1"
 }